Initial code cleanup and restructuring
