+++ bayesflow/networks/helpers/__init__.py
@@ -1,4 +1,4 @@
-from .blocks import DenseBlock, ConditionalDenseBlock, FFN
+from .blocks import DenseBlock, ConditionalDenseBlock, FFN, TimeTransformerBlock
 from .embeddings import FourierEmbedding, Time2Vec, RecurrentEmbedding, FiLM
 from .sequential import Sequential
 from .standardization import Standardization
