+++ bayesflow/networks/helpers/blocks/__init__.py
@@ -1,3 +1,4 @@
 from .dense_block import DenseBlock
 from .conditional_dense_block import ConditionalDenseBlock
-from .feedforward_for_tranformer import FFN
+from .transformer_feedforward import FFN
+from .time_transformer_block import TimeTransformerBlock
