+++ bayesflow/networks/helpers/blocks/transformer_feedforward.py
