+++ bayesflow/networks/subnets/transformer/__init__.py
@@ -1,0 +1,1 @@
+from .time_transformer import TimeTransformer
